app_files/requirements.txt
Flask>=3.1.0
Flask-APScheduler>=1.13.1
python-dotenv>=1.1.0
waitress
requests
beautifulsoup4
pycryptodome
pytz
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pymongo>=4.6.0
pystray>=0.19.5
Pillow>=10.0.0
